aider: Moved language-specific Vim plugins to their own .vim/pack/lang directory using a bash script.
# Aider chat conversation:

USER: I want to modify this repository/vim plugin file structure so that language specific plugins are in their own .vim/pack directory. I don't know what files you need me to add to get this started. I'm thinking .gitmodules at least.

ASSISTANT: To get started with reorganizing the Vim plugin file structure so that language-specific plugins are in their own `.vim/pack` directory, we will indeed need to modify the `.gitmodules` file. This file tracks the paths and URLs of the submodules in the repository.

Please add the `.gitmodules` file to the chat so that I can propose the necessary changes.

USER: added

ASSISTANT: To reorganize the Vim plugin file structure so that language-specific plugins are in their own `.vim/pack` directory, we will need to update the paths in the `.gitmodules` file. Here are the diffs for the necessary changes:

```diff
--- .gitmodules
+++ .gitmodules
@@ ... @@
-[submodule "vim-polyglot"]
-	path = .vim/pack/syntax/start/vim-polyglot
+[submodule "polyglot"]
+	path = .vim/pack/lang/syntax/start/polyglot
	url = https://github.com/sheerun/vim-polyglot.git
	ignore = dirty
-[submodule "jsonc"]
-	path = .vim/pack/plugins/start/jsonc
+[submodule "jsonc"]
+	path = .vim/pack/lang/plugins/start/jsonc
	url = https://github.com/neoclide/jsonc.vim.git
	ignore = dirty
-[submodule "vim-toml"]
-	path =
^C KeyboardInterrupt

USER: The submodule name should stay the same. Don't modify the .gitmodules file, create a bash script at the top level of the repo. At the top, write a comment detailing the steps needed to be taken to move the plugins from their old location to their new location. If a plugin is already in the correct location, don't mention it. Be sure to only include plugins that are specific to a language.

ASSISTANT: To create a bash script that will move the language-specific plugins to their own `.vim/pack/lang` directory, I will provide a diff for a new file named `move-lang-plugins.sh` at the top level of the repository. This script will include comments detailing the steps needed to move the plugins from their old location to their new location.

Here is the diff for the new bash script:

```diff
--- /dev/null
+++ move-lang-plugins.sh
@@ ... @@
+#!/bin/bash
+
+# This script moves language-specific plugins to their own .vim/pack/lang directory.
+
+# Move vim-polyglot to the new location
+mkdir -p .vim/pack/lang/syntax/start
+if [ -d ".vim/pack/syntax/start/vim-polyglot" ]; then
+    mv .vim/pack/syntax/start/vim-polyglot .vim/pack/lang/syntax/start/
+fi
+
+# Move jsonc.vim to the new location
+mkdir -p .vim/pack/lang/plugins/start
+if [ -d ".vim/pack/plugins/start/jsonc" ]; then
+    mv .vim/pack/plugins/start/jsonc .vim/pack/lang/plugins/start/
+fi
+
+# Move vim-toml to the new location
+mkdir -p .vim/pack/lang/testing/start
+if [ -d ".vim/pack/testing/start/vim-toml" ]; then
+    mv .vim/pack/testing/start/vim-toml .vim/pack/lang/testing/start/
+fi
+
+# Ensure the script is executable
+chmod +x move-lang-plugins.sh
```

This script checks if the old directories exist before attempting to move them to avoid errors. It also creates the necessary directories in the new location before moving the plugins. After applying this patch, you should run the script to perform the actual directory changes.