aider: Added vint as an ALE linter for Vimscript in the ftplugin/vim.vim file.
# Aider chat conversation:

USER: Are there any coc.vim settings needed for vim-language-server?

ASSISTANT: To configure `vim-language-server` for use with `coc.nvim`, you would typically add the language server configuration to the `coc-settings.json` file. Here is the diff that adds the `vim-language-server` configuration to the `coc-settings.json`:

```diff
--- .vim/coc-settings.json
+++ .vim/coc-settings.json
@@ ... @@
   "languageserver": {
+    "vimls": {
+      "command": "vim-language-server",
+      "args": ["--stdio"],
+      "filetypes": ["vim"],
+      "initializationOptions":
^C KeyboardInterrupt

USER: Are there any coc.vim settings needed for vim-language-server? Don't make any changes, the coc-settings file already has those settings.

ASSISTANT: If the `coc-settings.json` file already contains the necessary settings for `vim-language-server`, then no changes are needed. The `coc.vim` settings for `vim-language-server` are already in place and should be functioning as expected. If you encounter any issues or require further assistance with the configuration, feel free to ask.

USER: Remove any settings in the coc-settings vimlsp language server setting that is using default values.

USER: What settings are needed for vint? ale, coc, plugin? Don't change anything.

ASSISTANT: To integrate `vint` for linting Vimscript files, you can configure it in ALE, CoC, or as a standalone plugin. Here's a summary of what settings would be needed for each:

### ALE
For ALE, you would add `vint` to the list of linters for Vimscript files in the ALE configuration file, typically located at `.vim/pack/settings/start/settings/plugin/ale.vim`. You would also ensure that `vint` is installed on your system.

### CoC
For CoC, you would need to install a CoC extension that supports Vimscript linting, if available. If there's a `coc-vint` or similar extension, you would add it to the `g:coc_global_extensions` list in your `.vimrc`. You would also configure the `coc-settings.json` to use `vint` as a linter for Vimscript files.

### Standalone Plugin
If you were to use `vint` as a standalone plugin, you would install the plugin according to its documentation and configure it as needed, typically in a Vimscript file dedicated to `vint` settings.

Since you've asked not to make any changes, I will not provide diffs for these configurations. If you decide you want to implement any of these settings, please let me know, and I can provide the necessary diffs.

USER: Take into account I'm using both ale and coc, which is the preferred configuration? If a vim file needs to be changed, use the settings/ftplugin/vim.vim file, otherwise use the coc-settings file.

ASSISTANT: Given that you are using both ALE and CoC, the preferred configuration would be to let CoC handle the language server integration and completion, while ALE can be used for linting with tools that are not covered by CoC or for which you prefer ALE's handling.

For `vint`, since it is a linter and not a language server, it would be more appropriate to configure it with ALE. You would ensure that `vint` is installed on your system and then configure ALE to use `vint` for Vimscript files.

Here is the diff to add `vint` to the ALE linters for Vimscript in the `settings/ftplugin/vim.vim` file:

```diff
--- .vim/pack/settings/start/settings/ftplugin/vim.vim
+++ .vim/pack/settings/start/settings/ftplugin/vim.vim
@@ ... @@
 " Configure ALE to use vim-language-server for Vimscript
 let b:ale_linters = ['vimls']
 let b:ale_fixers = ['vimls']

 let b:ale_vim_vimls_executable = 'vim-language-server'
 let b:ale_vim_vimls_use_global = 1
+
+" Add vint as an ALE linter for Vimscript
+let b:ale_linters['vim'] = ['vint']
```

This configuration assumes that `vint` is already installed on your system. If it is not, you would need to install it using your system's package manager or via pip with `pip install vim-vint`.

For CoC, no changes are needed in the `coc-settings.json` file regarding `vint`, as it does not provide language server functionality, and CoC's linting capabilities for Vimscript will be deferred to ALE.