aider: Fix sed command issue in bin/update-coc-extensions script by updating syntax and approach for replacing lines.
# Aider chat conversation:

USER: I ran this command:

 bin/update-coc-extensions

And got this output:

sed: -e expression #1, char 86: unknown command: `''

ASSISTANT: It seems like there's an issue with the `sed` command in the `bin/update-coc-extensions` script. The error message indicates that `sed` is encountering an unknown command, which is likely due to an incorrect `sed` expression or syntax error in the script.

To resolve this issue, I will need to examine the `sed` command in the script and correct the syntax. Please provide the contents of the `bin/update-coc-extensions` script, or add it to the chat so I can make the necessary changes.

USER: it's in the chat

ASSISTANT: ```diff
--- bin/update-coc-extensions
+++ bin/update-coc-extensions
@@ ... @@
 #!/usr/bin/env bash

 ## script that creates a file `install-coc-extensions.vim`
 ## Use this command to install it (coc.nvim plugin required):
 ##        vim -c 'source install-coc-extensions.vim'

-EXTENSIONS=$(jq '.dependencies | keys' ~/.config/coc/extensions/package.json | sed 's/"/'"'"'/g;s/^/  \\/g')
+EXTENSIONS=$(jq -r '.dependencies | keys | .[]' ~/.config/coc/extensions/package.json | sed "s/.*/'&',/")

 echo "let g:coc_global_extensions = [" > install-coc-extensions.vim
 echo "$EXTENSIONS" >> install-coc-extensions.vim
 echo "]" >> install-coc-extensions.vim

-# Update .vimrc with the new list of extensions
-sed -i '/let g:coc_global_extensions = \[/,/\]/c\
-let g:coc_global_extensions = [\
-'"$EXTENSIONS"'\
-]' .vimrc
+# Update .vimrc with the new list of extensions, ensuring proper escaping of the sed delimiter
+sed -i.bak -e "/let g:coc_global_extensions = \[/,/\]/c\let g:coc_global_extensions = [\n$EXTENSIONS\n]" .vimrc
```

This diff should fix the `sed` command issue in the `bin/update-coc-extensions` script. The changes include using `jq` with the `-r` flag to output raw strings, thus avoiding the need for complex escaping in `sed`. Additionally, the `sed` command has been updated to use a different approach for replacing the lines between the brackets, which should resolve the error you encountered.