aider: Move the `ShowLanguageServer` function and the `LanguageServer` command to the `coc.vim` settings file.
# Aider chat conversation:

USER: Move that function to the coc.vim settings file

ASSISTANT: To move the `ShowLanguageServer` function and the `LanguageServer` command from `.vimrc` to `.vim/pack/settings/start/settings/plugin/coc.vim`, I will provide two diffs: one to remove the function and command from `.vimrc`, and another to add them to `coc.vim`.

Here is the diff to remove the function and command from `.vimrc`:

```diff
--- .vimrc
+++ .vimrc
@@ ... @@
 " Compile the help tags for all plugins
 helptags ALL

-function! ShowLanguageServer()
-  let l:info = coc#rpc#request('workspace/showOutput', 'languageserver')
-  if l:info == v:null
-    echo "No language server is active for the current buffer."
-  else
-    echo l:info
-  endif
-endfunction
-
-command! LanguageServer call ShowLanguageServer()

 colorscheme norwaytoday
```

And here is the diff to add the function and command to `coc.vim`:

```diff
--- .vim/pack/settings/start/settings/plugin/coc.vim
+++ .vim/pack/settings/start/settings/plugin/coc.vim
@@ ... @@
 "nmap <leader>qf  <Plug>(coc-fix-current)
+
+function! ShowLanguageServer()
+  let l:info = coc#rpc#request('workspace/showOutput', 'languageserver')
+  if l:info == v:null
+    echo "No language server is active for the current buffer."
+  else
+    echo l:info
+  endif
+endfunction
+
+command! LanguageServer call ShowLanguageServer()
```

These diffs will move the `ShowLanguageServer` function and the `LanguageServer` command from `.vimrc` to the `coc.vim` settings file.