aider: Updated ShowLanguageServer function to display active language servers using CocList services command.
# Aider chat conversation:

USER: Running LanguageServer command is giving this error: Error detected while processing function ShowLanguageServer[1]..coc#rpc#request[4]..<SNR>31_request[34]..function ShowLanguageServer[1]..coc#rpc#request[4]..<S
NR>31_request:
line   29:
E605: Exception not caught: Error on "workspace/showOutput" request: Action "workspace/showOutput" not exist

ASSISTANT: It seems that the `workspace/showOutput` action does not exist in the version of CoC you are using. To resolve this issue, we need to modify the `ShowLanguageServer` function to use a different approach to retrieve the language server information.

We can use the `:CocList services` command to show the running services, which includes language servers. We can then create a custom command that will display the output of this command filtered to show only the relevant information for the current buffer.

Here is the diff to update the `ShowLanguageServer` function in the `coc.vim` settings file:

```diff
--- .vim/pack/settings/start/settings/plugin/coc.vim
+++ .vim/pack/settings/start/settings/plugin/coc.vim
@@ ... @@
 "nmap <leader>qf  <Plug>(coc-fix-current)

-function! ShowLanguageServer()
-  let l:info = coc#rpc#request('workspace/showOutput', 'languageserver')
-  if l:info == v:null
-    echo "No language server is active for the current buffer."
-  else
-    echo l:info
-  endif
-endfunction
+function! ShowLanguageServer() abort
+  redir => l:output
+  silent CocList services
+  redir END
+  echo l:output
+endfunction

 command! LanguageServer call ShowLanguageServer()
```

This updated function will redirect the output of `:CocList services` to a variable and then print it, which should show the active language servers for the current buffer.