aider: Ensure `UPDATE_EXTENSIONS` string is treated as a literal during substitution in `bin/update-coc-extensions` script.
# Aider chat conversation:

USER: What is \Q in a perl regex, and what is the correct option to use here to escape any special character in the string?

ASSISTANT: In Perl regular expressions, `\Q` is used to quote (escape) all the characters up to `\E`. This means that any characters that would normally be treated as special characters in a regex are treated as literal characters instead.

However, the `\Q...\E` construct is not necessary when you want to replace a literal string in Perl, as the `s///` operator can handle that without needing to escape each character individually. If you want to escape special characters in a string that is being used as part of a regex pattern, you can use the `\Q...\E` construct.

In the context of the `bin/update-coc-extensions` script, if you want to ensure that any special characters in the `UPDATE_EXTENSIONS` variable are treated as literals when performing the substitution, you would use `\Q` before the variable and `\E` after it in the regex pattern.

Here's the correct usage in the context of the script:

```diff
--- bin/update-coc-extensions
+++ bin/update-coc-extensions
@@ ... @@
-perl -i -pe "BEGIN{undef $/;} s#let g:coc_global_extensions = \[.*?\]#\Q$UPDATE_EXTENSIONS#s" .vimrc
+perl -i -pe "BEGIN{undef $/;} s#let g:coc_global_extensions = \[.*?\]#\Q$UPDATE_EXTENSIONS\E#s" .vimrc
```

This change ensures that the `UPDATE_EXTENSIONS` string is treated as a literal string during the substitution, and any special characters within it will not be interpreted as regex metacharacters.